+++ packages/auth/CHANGELOG.md
@@ -1,3 +1,12 @@
+## @lunora/auth [1.0.0-alpha.90](https://github.com/anolilab/lunora/compare/@lunora/auth@1.0.0-alpha.89...@lunora/auth@1.0.0-alpha.90) (2026-08-25)
+
+## @lunora/auth [1.0.0-alpha.89](https://github.com/anolilab/lunora/compare/@lunora/auth@1.0.0-alpha.88...@lunora/auth@1.0.0-alpha.89) (2026-08-24)
+
+
+### Dependencies
+
+* **@lunora/server:** upgraded to 1.0.0-alpha.80
+
 ## @lunora/auth [1.0.0-alpha.88](https://github.com/anolilab/lunora/compare/@lunora/auth@1.0.0-alpha.87...@lunora/auth@1.0.0-alpha.88) (2026-08-24)
 
 ### Features